Prevent users from removing their last login if they don't have a password
While this is checked for in `GET /ExternalLogins` (by not showing the remove button), this is not confirmed in `POST /RemoveLogin`. This means a malicious user without a password can manually send a POST request to `/RemoveLogin` containing the details of their last login! This would result in having a password-less user in the database with no external logins.