Added support for query string parameters
I attempted to register a route that had a query string parameter but crossroads wasn't picking it up.  Here is an example route.  
```
{ url: 'linkaccount{?query}', params: { page: 'streaming-services' } },
```
The problem was that the query string portion of the url was not included when invoking `crossroads.parse()`.  By including `document.location.search` in the string being passed to `crossroads.parse()`, this issue is resolved.  Followed the History.js example from the Crossroads documentation.  https://github.com/millermedeiros/crossroads.js/wiki/Examples#using-with-historyjs